Create GZ raw data veiw
Maybe, since I will never ever be suggesting code, you would prefer me to just describe the feature or bug here.

When you Examine data in GZ and choose the option to view the raw data, you are allowed to choose the filter you want to view.  
After you select a filter, it would be helpful if that button remained shaded.